+++ index.html
@@ -49,7 +49,7 @@
       gtag("config", "G-XZZPDDGCRR");
     </script>
 
-    <script src="/src/js/script.js" type="module"></script>
+    <script src="/src/main.js" type="module"></script>
   </head>
 
   <body>
