+++ src/main.js
@@ -1,0 +1,15 @@
+/**
+ * script.js
+ *
+ * @license MIT
+ *
+ */
+
+// eslint-disable-next-line
+import './styles/style.scss';
+
+import { menu, projects, contact } from './modules/index.js';
+
+menu();
+projects();
+contact();
