+++ src/modules/index.js
@@ -1,0 +1,3 @@
+export { default as contact } from './contact.js';
+export { default as menu } from './menu.js';
+export { default as projects } from './projects.js';
