+++ src/styles/sections/_about.scss
