+++ src/styles/sections/_footer.scss
