+++ src/styles/sections/_modal.scss
