+++ src/styles/sections/_projects.scss
