+++ src/styles/sections/_who.scss
