+++ vite.config.js
@@ -3,6 +3,11 @@
  */
 const config = {
   publicPath: process.env.NODE_ENV === 'production' ? '/portfolio' : '/',
+  resolve: {
+    alias: {
+      '@': './src',
+    },
+  },
 };
 
 export default config;
